Rename to use standard revenuecat tester package and bundle id
