+++ examples/purchaseTesterTypescript/android/app/build.gradle
@@ -131,7 +131,7 @@ android {
     compileSdkVersion rootProject.ext.compileSdkVersion
 
     defaultConfig {
-        applicationId "com.revenuecat.purchase_tester"
+        applicationId "com.revenuecat.purchases_sample"
         minSdkVersion rootProject.ext.minSdkVersion
         targetSdkVersion rootProject.ext.targetSdkVersion
         versionCode 1
