+++ examples/purchaseTesterTypescript/android/app/src/main/java/com/revenuecat/purchases_sample/MainApplication.java
@@ -1,4 +1,4 @@
-package com.revenuecat.purchase_tester;
+package com.revenuecat.purchases_sample;
 
 import android.app.Application;
 import android.content.Context;
@@ -62,7 +62,7 @@ public class MainApplication extends Application implements ReactApplication {
          We use reflection here to pick up the class that initializes Flipper,
         since Flipper library is not available in release mode
         */
-        Class<?> aClass = Class.forName("com.purchase_tester.ReactNativeFlipper");
+        Class<?> aClass = Class.forName("com.revenuecat.purchases_sample.ReactNativeFlipper");
         aClass
             .getMethod("initializeFlipper", Context.class, ReactInstanceManager.class)
             .invoke(null, context, reactInstanceManager);
